Refactors IntervalPersistor tests to support EventPersistor tests
Refs #2
