+++ WatchDogEclipsePlugin/WatchDogUnitTests/src/nl/tudelft/watchdog/logic/interval/IntervalPersisterInvalidDatabaseTest.java
@@ -12,7 +12,7 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class IntervalPersisterInvalidDatabaseTest extends PersisterTestBase {
+public class IntervalPersisterInvalidDatabaseTest extends IntervalPersisterTestBase {
 
 	@BeforeClass
 	public static void setUpBeforeClass() throws FileNotFoundException,
