+++ WatchDogEclipsePlugin/WatchDogUnitTests/src/nl/tudelft/watchdog/logic/storage/LexicographicalTestOrderRunner.java
@@ -1,4 +1,4 @@
-package nl.tudelft.watchdog.logic.interval;
+package nl.tudelft.watchdog.logic.storage;
 
 import java.util.ArrayList;
 import java.util.Collections;
